Fixes variable parsing with lowercase letters, Bump version
Previously when sing `devn tf prod` variables in the format
TF_VAR_foo=bar would be ignored and would not be exported nor
tracked by denv. This is a bug in load-env.
See here for details: https://github.com/pbrisbin/load-env/issues/3

Until this is fixed and released I'm using a fork