Adds support for setting GOOGLE_CREDENTIALS
This is useful when talking to GCP services. Usually I need it for
using terraform or various CLI tooling that expects
`GOOGLE_CREDENTIALS=/path/to/service_account.json` to be set.

`denv gcp -p ~/.gcp/prod-admin.json` will do just that now.

In the future we might want to add some sort of temporary credentials
like we do with AWS.